Allow force-recreate on certificate
Previous logic did not re-run if a certificate expired. The new logic,
is pretty basic, but basically allow a variable set that forces
generation of a new signed certificate + private key. Probably could
have done this in a smarter way to retain same private key.. but :shrug:
I got things to do!